Update 1.0.1
I figured out that my initial issue with a nullPointerException was coming from my deposit.jsp and  withdraw.jsp files becuase I was referencing the servlet twice in my form so one of the variables was being reset to null. This was causing my method to fail for referencing a null variable.